-Added Summer 2012 anime season data. -Japanese counters (e.g. "第12話") in file names are now also recognized. -The image in anime information dialog is now only refreshed when necessary. -Proxy settings are now used in season dialog as well. -Fixed a bug which caused a series' airing status to be displayed incorrectly.
git-svn-id: http://taiga.googlecode.com/svn/trunk@165 77add795-4bdf-f0f6-a574-f4d068491005

+++ time.cpp
@@ -35,7 +35,7 @@ Date::Date(const wstring& date) {
   *this = mal::ParseDateString(date);
 }
 
-Date::Date(int year, int month, int day)
+Date::Date(unsigned short year, unsigned short month, unsigned short day)
     : year(year), 
       month(month), 
       day(day) {
